feat(storage): atomic KvStore::compare_and_swap; split kv.rs into submodules
Closes the kv_cas atomicity gap flagged by Gemini's PR #752 review.

KvStore trait
-------------
- New `async fn compare_and_swap(&self, namespace, key, expected, new)
  -> StorageResult<bool>`. Returns Ok(true) when the swap landed,
  Ok(false) when the predicate didn't match (now or by commit time),
  Err only for I/O / validation failures. `expected = None` means
  insert-if-absent.

MemoryKvStore
-------------
- Single write lock covers the entire read+conditional-write — other
  capsules calling set / compare_and_swap on the same store block
  until the CAS returns, so the compare cannot race a concurrent
  mutation.

SurrealKvStore
--------------
- Backend-level `tokio::sync::Mutex<()>` guards the full
  begin+read+set+commit sequence. SurrealKV's
  Transaction::validate_write_conflicts reads the memtable *before*
  Core::commit acquires its write mutex, so two concurrent commits
  can both pass validation and both succeed — observable as a
  multi-winner race in the new concurrent test.
- Conflict-detection variant matched against
  surrealkv::Error::TransactionWriteConflict /
  ::TransactionRetry (no more string scraping that a future error
  message rewording could silently break).

ScopedKvStore
-------------
- New `compare_and_swap(key, expected, new)` method delegating to
  the inner store under the bound namespace.

capsule host
------------
- kv_cas now calls kv.compare_and_swap directly; the get-then-set
  emulation + TODO comment removed.

Tests
-----
- Per-impl tests for the basic cases (insert-if-absent matches /
  doesn't match, replace when expected matches / differs, missing
  key + Some(expected) → false).
- Multi-thread tokio runtime concurrent test on both backends: 16-32
  tasks racing CAS(Some(b"0") → "winner-{i}") on the same key.
  Exactly one must win. (Required prefixing per-task values so a task
  writing "0" can't leave the store in its initial state.)

Refactor
--------
- kv.rs was 1226 lines after the CAS addition — over the project's
  1000-line CI ceiling — so split into kv/mod.rs (validators, helpers,
  trait, re-exports) + kv/memory.rs + kv/surreal.rs + kv/scoped.rs.
  Public API preserved verbatim via pub use.
