+++ CHANGELOG.md
@@ -25,7 +25,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 - **Per-domain WIT review fixups round 2 (Gemini, PR #752).**
   - `MAX_ACTIVE_STREAMS` / `MAX_SUBSCRIPTIONS` / `MAX_BACKGROUND_PROCESSES` quota gates now read O(1) counter fields on `HostState` (`net_stream_count`, `subscription_count`, `process_count_total`, `process_count_by_principal`) instead of walking the entire `ResourceTable`. Each successful resource insert bumps the counter; the matching `drop` impl decrements. Per-principal sub-budgets for `spawn-background` use a `HashMap<PrincipalId, usize>` keyed on the creator. Single-threaded: wasmtime stores are owned by exactly one OS thread.
   - CI workflows (`ci.yml`) now check out the `wit/` submodule recursively so `astrid-capsule`'s build script can stage the per-domain WIT packages. Previously every job that touched the build (`check`, `clippy`, `test`, `msrv`) panicked with `read wit/host: No such file or directory`.
-  - Note: `kv_cas` atomicity gap (`get` → `set` not atomic across capsules) remains a known TODO. The fix requires plumbing a `compare_and_swap` primitive through `KvStore` / `ScopedKvStore` in `astrid-storage`, which is out of scope for the WIT split PR. Tracking issue to follow.
+- **Atomic `kv_cas` across capsules.** The host fn used to emulate compare-and-swap with a `get` followed by a `set`, which raced across capsules running concurrently in the same kernel. `KvStore` now exposes an atomic `compare_and_swap(namespace, key, expected, new)` primitive; `MemoryKvStore` serializes read+conditional-write under its existing `RwLock`, and `SurrealKvStore` runs the comparison inside a single MVCC transaction guarded by an internal mutex that closes the TOCTOU between surrealkv's `validate_write_conflicts` and its actual write phase. Two concurrent-spawn regression tests (memory + surreal) assert exactly one of 16–32 racing tasks succeeds the swap. The capsule `kv_cas` host fn delegates straight to the new primitive — no more emulation.
+- **`astrid-storage::kv` split into a submodule directory.** `kv.rs` (now ~1200 lines after the CAS addition) became `kv/mod.rs` (validators, helpers, trait, re-exports) + `kv/memory.rs` (`MemoryKvStore`) + `kv/surreal.rs` (`SurrealKvStore`, behind the `kv` feature) + `kv/scoped.rs` (`ScopedKvStore`), each well under the 1000-line CI ceiling. Public API preserved verbatim via `pub use`.
 
 ### Added
